requirements.txt
keyboard==0.13.5
Pillow==8.0.1
pynput==1.7.1
pyperclip==1.8.1
pytesseract==0.3.6
six==1.15.0
